// src/interfaces/interceptor.interface.ts
export interface IHeadersCore {
  set(name: string, value: string | string[]): void | unknown;
  append(name: string, value: string | string[]): void | unknown;
  // keys(): string[] | Iterator<string>;
  get(name: string): string | string[] | null;
  // getAll?(name: string): string | string[] | null;
}

export interface IRequestCore<T> {
  method: string;
  url: string;
  urlWithParams?: string;
  headers?: IHeadersCore;
  body?: T;
}

export interface IResponseBase {

  /**
   * Response headers
   */
  headers?: IHeadersCore;

  /**
   * Http {@link http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html status code}
   * associated with the response.
   */
  status?: number;

  /**
   * Status text for the status code
   */
  statusText?: string;
  /**
   * request url
   */
  url?: string;
}

export interface IHttpResponse<T> extends IResponseBase {
  body?: T | null;
}


export interface IErrorMessage {
  message: string;
  detailedMessage?: string | unknown;
}

export interface IHttpErrorResponse extends IResponseBase {
  error?: IErrorMessage | unknown;
}
// ...
export interface IPassThruBackend {
  /**
   * Handle an HTTP request and return an Observable of HTTP response
   * Both the request type and the response type are determined by the supporting HTTP library.
   */
  handle(req: IRequestCore<unknown>): Promise<IHttpResponse<unknown> | IHttpErrorResponse>;
}